chore:Removed Dupicate AcceptanceTestBase (#7815)
* chore:Removed Dupicate AcceptanceTestBase


* formatting


---------
