+++ acceptance-tests/tests/src/test/java/org/hyperledger/besu/tests/acceptance/clique/CliqueZeroValidatorsAcceptanceTest.java
@@ -14,14 +14,14 @@
  */
 package org.hyperledger.besu.tests.acceptance.clique;
 
-import org.hyperledger.besu.tests.acceptance.dsl.AcceptanceTestBaseJunit5;
+import org.hyperledger.besu.tests.acceptance.dsl.AcceptanceTestBase;
 import org.hyperledger.besu.tests.acceptance.dsl.node.BesuNode;
 
 import java.io.IOException;
 
 import org.junit.jupiter.api.Test;
 
-public class CliqueZeroValidatorsAcceptanceTest extends AcceptanceTestBaseJunit5 {
+public class CliqueZeroValidatorsAcceptanceTest extends AcceptanceTestBase {
 
   @Test
   public void zeroValidatorsFormValidCluster() throws IOException {
